This is what I've been working with.  Most changes are in.

+++ src/r_main.h
@@ -52,7 +52,7 @@ extern fixed_t		centerxfrac;
 extern fixed_t		centeryfrac;
 extern fixed_t		projection;
 
-extern int		validcount;
+extern short		validcount;
 
 extern int		linecount;
 extern int		loopcount;
